+++ tests/general/tools/test_delete.py
@@ -232,6 +232,8 @@ _PRB_MINIMAL_BODY = textwrap.dedent(
     """\
     # Simple Problem Statement
 
+    The current process is causing delays, for users because of missing automation.
+
     ## Current State
 
     ### Summary